+++ src/pytest_mock/__init__.py
@@ -1,4 +1,6 @@
+from pytest_mock.plugin import AsyncMockType
 from pytest_mock.plugin import MockerFixture
+from pytest_mock.plugin import MockType
 from pytest_mock.plugin import PytestMockWarning
 from pytest_mock.plugin import class_mocker
 from pytest_mock.plugin import mocker
@@ -11,8 +13,10 @@ from pytest_mock.plugin import session_mocker
 MockFixture = MockerFixture  # backward-compatibility only (#204)
 
 __all__ = [
+    "AsyncMockType",
     "MockerFixture",
     "MockFixture",
+    "MockType",
     "PytestMockWarning",
     "pytest_addoption",
     "pytest_configure",