+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Releases
 ========
 
+UNRELEASED
+----------
+
+* `#415 <https://github.com/pytest-dev/pytest-mock/pull/415>`_: ``MockType`` and ``AsyncMockType`` can be imported from ``pytest_mock`` for type annotation purposes.
+
+
 3.13.0 (2024-03-21)
 -------------------